+++ src/App.vue
@@ -5,21 +5,22 @@
 
 <script setup>
 import { RouterView } from "vue-router";
-
 import { onMounted } from "vue";
+
 import { useAccountStore } from "@/stores/account";
-import { IDBgettable, decryptText } from "@/idb";
+import { useMapStore } from "@/stores/map";
+
+import { decryptText } from "@/idb";
 import config from "@/config/default/config.json";
+import { getSensorsForLastDay } from "./utils/utils";
 
 const accountStore = useAccountStore();
+const mapStore = useMapStore();
 
 onMounted(async () => {
 
-  const accounts = await IDBgettable(
-    config.INDEXEDDB.accounts.dbname,
-    config.INDEXEDDB.accounts.dbversion,
-    config.INDEXEDDB.accounts.tablename
-  );
+  /* + INIT ACCOUNT */
+  const accounts = await accountStore.getDB();
 
   if (accounts && accounts.length > 0) {
     let selected = accounts.find(acc => acc.active);
@@ -39,6 +40,15 @@ onMounted(async () => {
       selected.devices
     );
   }
+  /* - INIT ACCOUNT */
+
+  /* + INIT SENSORS */
+  const sensors = await getSensorsForLastDay();
+  if (Array.isArray(sensors)) {
+    mapStore.setSensors(sensors);
+  }
+  /* - INIT SENSORS */
+
 });
 </script>