+++ src/utils/map/marker.js
@@ -323,12 +323,6 @@ export async function addPoint(point) {
         console.log(point);
       } else if (point.model === 2) {
         await addMarker(point);
-
-        /* + Broadcasting about new sensor */
-        const bc = new BroadcastChannel("sensors");
-        bc.postMessage(point);
-        bc.close();
-        /* - Broadcasting about new sensor */
       } else if (point.model === 3) {
         await addMarker(point);
         await addPointPath(point);
@@ -511,12 +505,6 @@ async function addMarkerUser(point) {
 export function clear() {
   if (markersLayer) {
     markersLayer.clearLayers();
-
-    /* + Broadcasting about removing all sensors */
-    const bc = new BroadcastChannel("sensorsremoved");
-    bc.postMessage(true);
-    bc.close();
-    /* - Broadcasting about removing all sensors */
   }
   if (pathsLayer) {
     pathsLayer.clearLayers();